Old implementation is commited here that does not work. Saved for future changes in case if needed. In next commit will be migrated to Hibernate Repositories (Stateless Session implementation)

+++ src/main/java/org/tuvarna/chat/model/write/command/ChatMessageCommand.java
@@ -1,8 +1,5 @@
 package org.tuvarna.chat.model.write.command;
 
-import org.tuvarna.chat.model.write.command.ChatMessageCommand.ArchiveMessage;
-import org.tuvarna.chat.model.write.command.ChatMessageCommand.SendMessages;
-import org.tuvarna.chat.model.write.command.ChatMessageCommand.UpdateMessage;
 import org.tuvarna.chat.model.write.dto.ChatMessageSaveData;
 
 import java.util.List;
@@ -15,10 +12,10 @@ public sealed interface ChatMessageCommand permits
     record SendMessages(List<ChatMessageSaveData> requestList) implements ChatMessageCommand {
     }
 
-    record ArchiveMessage(int messageId) implements ChatMessageCommand {
+    record ArchiveMessage(long messageId) implements ChatMessageCommand {
     }
 
-    record UpdateMessage(int messageId, String updatedContent) implements ChatMessageCommand {
+    record UpdateMessage(long messageId, String updatedContent) implements ChatMessageCommand {
     }
 
 }
